+++ src/advantage/formats/index.ts
@@ -1,5 +1,13 @@
 import { topscroll } from "./topscroll";
 import { midscroll } from "./midscroll";
 import { welcomePage } from "./welcomepage";
+import { doubleMidscroll } from "./double-midscroll";
+import { tripleMidscroll } from "./triple-midscroll";
 
-export const defaultFormats = [topscroll, midscroll, welcomePage];
+export const defaultFormats = [
+    topscroll,
+    midscroll,
+    welcomePage,
+    doubleMidscroll,
+    tripleMidscroll
+];